build: unblock all-targets assemble — bump AGP, isolate JVM diags, fix test name
Three changes that together let `./gradlew assemble` succeed across the
full multiplatform target matrix (jvm, android, linuxX64, linuxArm64,
iosArm64, iosSimulatorArm64, macosArm64, js, wasmJs, wasmWasi):

1. **agp 9.1.1 → 9.2.0** in libs.versions.toml. Composite-build SKaiNET
   uses 9.2.0; the version skew was triggering "Using multiple versions
   of the Android Gradle plugin in the same build is not allowed".

2. **Isolate JVM-only diagnostic code via `expect/actual`.**
   `HybridTransformerBlock`, `MultiHeadAttention`, and `GemmaModel` all
   had `System.getenv("GEMMA4_DUMP_*")` gates and tensor-dump bodies in
   commonMain that referenced JVM-only APIs (`System`, `String.format`,
   `MemorySegmentTensorData`, `java.lang.foreign.MemorySegment`). The
   metadata compile for non-JVM targets failed on those.

   New shim in `sk.ainet.apps.llm.diag` (llm-core):
     - `expect fun envFlag(name: String): Boolean`
       JVM: `System.getenv(name) == "1"`
       Other targets (registryBasedMain): always false
     - `expect fun dumpStats(label: String, tensor: Tensor<*, *>)`
       JVM: full formatted dump (min/max/mean/rms + last-pos fingerprint)
       Other targets: no-op
   Both `HybridTransformerBlock` (DUMP_INNER, DUMP_MHA, DUMP_BLOCKS) and
   `GemmaModel` (DUMP_HIDDEN) now route through these. Body of
   `MultiHeadAttention.mhaDumpStat` collapses to a single call.
   `MultiHeadAttentionDiag.shouldDumpThisCall` lost `@JvmStatic`
   (commonMain doesn't see it; the field-access pattern doesn't need
   the JVM-static optimization).

   Diagnostic capability fully preserved on JVM (the only platform that
   ever ran them); permanently no-op on JS/wasm/native/Android (no JFR
   profiling story there worth maintaining).

3. **Rename `TokenizerEncodingTest` test method**: `()` is not a legal
   character in JS/wasm test names. The renamed method tests the same
   USER_DEFINED-tokens-emit-atomically behavior.

Combined with upstream `SKaiNET / 52d628c2` (which reverts the op-level
parallelism that pulled `kotlinx.coroutines.runBlocking` into commonMain
where it isn't multiplatform-resolvable), `./gradlew assemble` now
completes successfully across all targets.

JVM tests (llm-core, llm-agent, llm-inference:gemma) still green —
diagnostic plumbing change is functionally a no-op when the env vars
aren't set.
