+++ gradle/libs.versions.toml
@@ -1,6 +1,6 @@
 [versions]
 skainet = "0.20.0"
-agp = "9.1.1"
+agp = "9.2.0"
 jacksonDatabind = "2.21.2"
 jsonSchemaValidator = "3.0.1"
 jsonSchemaValidatorVersion = "0.5.4"
